+++ Tests/MapboxMapsTests/Annotations/ViewAnnotationTests.swift
@@ -215,48 +215,6 @@ final class ViewAnnotationTests: XCTestCase {
         XCTAssertEqual(anchorCoordinates, [descriotor1.anchorCoordinate, descriotor2.anchorCoordinate])
         XCTAssertEqual(visibilities, [true])
     }
-
-    /// Marker Tests
-    func testSettingMarkerProperties() {
-        let marker = Marker(coordinate: .testConstantValue())
-            .color(.blue)
-            .innerColor(.orange)
-            .stroke(.green)
-            .text(.testConstantValue())
-
-        XCTAssertEqual(marker.innerColor, .orange)
-        XCTAssertEqual(marker.outerColor, .blue)
-        XCTAssertEqual(marker.strokeColor, .green)
-        XCTAssertEqual(marker.coordinate, CLLocationCoordinate2D.testConstantValue())
-        XCTAssertEqual(marker.text, String.testConstantValue())
-    }
-
-    func testDefaultMarkerProperties() {
-        let marker = Marker(coordinate: .testConstantValue())
-
-        XCTAssertEqual(marker.innerColor, Color(red: 1, green: 1, blue: 1, opacity: 1.0))
-        XCTAssertEqual(marker.outerColor, Color(red: 207/255, green: 218/255, blue: 247/255, opacity: 1.0))
-        XCTAssertEqual(marker.strokeColor, Color(red: 58/255, green: 89/255, blue: 250/255, opacity: 1.0))
-        XCTAssertEqual(marker.coordinate, CLLocationCoordinate2D.testConstantValue())
-        XCTAssertEqual(marker.text, nil)
-        XCTAssertNil(marker.tapAction)
-    }
-
-    func testMarkerTapGesture() {
-        var tapped = false
-        let marker = Marker(coordinate: .testConstantValue())
-            .text("Test Marker")
-            .onTapGesture {
-                tapped = true
-            }
-
-        XCTAssertNotNil(marker.tapAction)
-        XCTAssertFalse(tapped)
-
-        // Execute the tap action
-        marker.tapAction?()
-        XCTAssertTrue(tapped)
-    }
 }
 
 class DummyAnnotationView: UIView {